refactor: move fundamental signals to core
