+++ packages/signal/src/core/event-signal.test.js
